#774.  Fix MSVC compiler warning.

+++ src/xmpsidecar.cpp
@@ -48,7 +48,7 @@ EXIV2_RCSID("@(#) $Id$")
 // *****************************************************************************
 namespace {
     const char* xmlHeader = "<?xpacket begin=\"\xef\xbb\xbf\" id=\"W5M0MpCehiHzreSzNTczkc9d\"?>\n";
-    const long  xmlHdrCnt = std::strlen(xmlHeader); // without the trailing 0-character
+    const long  xmlHdrCnt = (long) std::strlen(xmlHeader); // without the trailing 0-character
     const char* xmlFooter = "<?xpacket end=\"w\"?>";
 }
